Add make rules and linker script for the RP2040 stage2 bootloader
The RP2040 needs a special stage2 bootloader to allow starting it from flash.
The bootloader code is included in the Pico-SDK, but it has to be integrated
into the ChibiOS build system to be usable, because the code placement has
to be adapted. These files implement that. To use add the .mk files to the
project Makefile and use the RP2040_FLASH.ld linker script.

Also included is a .mk file to produce UF2 output files that allow to flash
the code with the USB bootloader of the RP2040.